feat(sidebar): show last-used provider icon on recent conversation rows
- Add shared ProviderBrandIcon component with isKnownProviderId guard
  (legacy rows persist "", optimistic rows may carry "pending" or a
  custom provider instance id — only enum hits render an icon)
- Render the icon left of the row title, hidden in selection mode;
  native title tooltip shows the last-used model name
- Compare providerId/model in areRenderedHistoryItemsEqual so rows
  re-render after a new turn persists a different model
- Converge two byte-identical private ProviderBrandIcon copies
  (ComposerModelControls, modelPicker) onto the shared component
